requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis~=5.0.7